Added THIRDPARTYDIR to build system
By default, thirdparty packages which are not included in the GR repository
are downloaded and installed to the 3rdparty/build directory. For OBS, these
are instead downloaded to build/3rdparty and should be installed there as well.
This commit adds the THIRDPARTYDIR and THIRDPARTYINCDIR variables to
lib/gks/Makefile, which controls where such dependencies should be installed
and might already be found. Previously this was only done via EXTRA_CFLAGS,
etc, but this variable allows the creation of Makefile targets for these
thirdparty packages and their files.

Note: This variable is **not** used by xcodebuild, so for macOS the default
      should still be used or the XCode project needs to be configured
      accordingly.